WIP IBaseSDK

+++ Siesa.SDK.Frontend/BusinessManager.cs
@@ -45,12 +45,12 @@ namespace Siesa.SDK.Frontend
             return Backend.ValidateAndSaveBusiness(Name, obj);
         }
 
-        public dynamic Delete(int id)
+        public dynamic Delete(Int64 id)
         {
             return Backend.DeleteBusinessObj(Name, id);
         }
 
-        public dynamic Get(int id)
+        public dynamic Get(Int64 id)
         {
             return Backend.GetBusinessObj(Name, id);
         }
