+++ Siesa.SDK.Frontend/Components/FormManager/Views/DynamicDetailView.razor.cs
@@ -8,7 +8,7 @@ namespace Siesa.SDK.Frontend.Components.FormManager.Views
 {
     public partial class DynamicDetailView : DynamicBaseViewModel
     {
-        private async Task InitDetail(int business_obj_id)
+        private async Task InitDetail(Int64 business_obj_id)
         {
             try
             {
@@ -36,7 +36,7 @@ namespace Siesa.SDK.Frontend.Components.FormManager.Views
                         //BusinessObj = null;
                         ErrorMsg = "";
 
-                        await InitDetail(Convert.ToInt32(BusinessObjId));
+                        await InitDetail(Convert.ToInt64(BusinessObjId));
                         StateHasChanged();
                     }
