+++ Siesa.SDK.Frontend/Components/FormManager/Views/DynamicEditView.razor.cs
@@ -8,7 +8,7 @@ namespace Siesa.SDK.Frontend.Components.FormManager.Views
 {
     public partial class DynamicEditView : DynamicBaseViewModel
     {
-        private async Task InitEdit(int business_obj_id){
+        private async Task InitEdit(Int64 business_obj_id){
             try
             {
                 await BusinessObj.InitializeBusiness(business_obj_id);
@@ -34,7 +34,7 @@ namespace Siesa.SDK.Frontend.Components.FormManager.Views
                         //BusinessObj = null;
                         ErrorMsg = "";
 
-                        await InitEdit(Convert.ToInt32(value));
+                        await InitEdit(Convert.ToInt64(value));
                         StateHasChanged();
                     }
